Fold SO3LR charge head into SimpleReadout+AtomicAffine; rename interaction.
Move HirshfeldReadout into readout.py and drop PartialChargeReadout in favor
of the equivalent SimpleReadout(Ea,Qa)+AtomicAffine(scale=1) stack. Rename
so3krates euclidean_transformer.py to interaction.py to match PhysNet/SpookyNet.
