+++ docs/user_guide/models/layer_stack.rst
@@ -35,7 +35,9 @@ From :code:`enerzyme/models/layers/`:
     :code:`EquiformerGraphAttentionReadout` — separate GraphAttention head over full
     irreps + graph edges (not mixed into SimpleReadout); use when you want an
     attention-style multi-field atomic scalar head.
-    :code:`PartialChargeReadout` / :code:`HirshfeldReadout` — SO3LR charge and Hirshfeld-ratio heads.
+    :code:`HirshfeldReadout` — SO3LR Hirshfeld-ratio head (``ha`` for TS–QDO).
+    Partial charges use :code:`SimpleReadout(Qa)` + :code:`AtomicAffine` with
+    fixed unit scale (element shift ≡ SO3LR ``Emb(Za)`` bias).
     Equiformer-series external readouts (including LinearRS / GraphAttention /
     :code:`EquiformerV2FeedForwardReadout`) accept :code:`shallow_ensemble_size`
     by widening the last linear layer.
@@ -171,7 +173,7 @@ SO3LR (So3krates + universal pairwise FF)
 
 :code:`architecture: so3lr` reuses :code:`So3kratesCore` and composes SO3LR-specific
 layers: :code:`ChargeSpinEmbedding` + :code:`GatherAtomEmbedding(scale_by_sqrt_count=true)`,
-:code:`PartialChargeReadout` / :code:`ChargeConservation` / :code:`HirshfeldReadout`,
+:code:`SimpleReadout(Qa)` / :code:`AtomicAffine` / :code:`ChargeConservation` / :code:`HirshfeldReadout`,
 then :code:`ZBLRepulsionEnergy(switch_off=1.5)`, :code:`ElectrostaticEnergy(flavor=SO3LR)`,
 and :code:`TSQDODispersionEnergy`. Do **not** substitute Grimme D3/D4 when targeting
 SO3LR dispersion. Example: :code:`enerzyme/config/so3lr_layers_example.yaml`. For
